chore: move PoolParams & related functions to a separate module.
