+++ crates/amaru-ledger/src/store/columns/pools.rs
@@ -226,100 +226,8 @@ impl<'a, C> cbor::decode::Decode<'a, C> for Row {
 #[cfg(any(test, feature = "test-utils"))]
 pub mod tests {
     use super::*;
-    use amaru_kernel::{
-        prop_cbor_roundtrip, Hash, IPv4, IPv6, Nullable, Port, RationalNumber, Relay,
-    };
-    use proptest::{collection, prelude::*};
-    use proptest::{collection::vec, prop_compose};
-
-    prop_compose! {
-        /// Generates arbitrary `PoolId` values using random 28-byte arrays.
-        pub fn any_pool_id()(
-            bytes in any::<[u8; 28]>(),
-        ) -> PoolId {
-            Hash::from(bytes)
-        }
-    }
-
-    fn any_nullable_port() -> impl Strategy<Value = Nullable<Port>> {
-        prop_oneof![
-            Just(Nullable::Undefined),
-            Just(Nullable::Null),
-            any::<u32>().prop_map(Nullable::Some),
-        ]
-    }
-
-    fn any_nullable_ipv4() -> impl Strategy<Value = Nullable<IPv4>> {
-        prop_oneof![
-            Just(Nullable::Undefined),
-            Just(Nullable::Null),
-            any::<[u8; 4]>().prop_map(|a| Nullable::Some(Vec::from(a).into())),
-        ]
-    }
-
-    fn any_nullable_ipv6() -> impl Strategy<Value = Nullable<IPv6>> {
-        prop_oneof![
-            Just(Nullable::Undefined),
-            Just(Nullable::Null),
-            any::<[u8; 16]>().prop_map(|a| Nullable::Some(Vec::from(a).into())),
-        ]
-    }
-
-    prop_compose! {
-        fn single_host_addr()(
-            port in any_nullable_port(),
-            ipv4 in any_nullable_ipv4(),
-            ipv6 in any_nullable_ipv6()
-        ) -> Relay {
-            Relay::SingleHostAddr(port, ipv4, ipv6)
-        }
-    }
-
-    prop_compose! {
-        fn single_host_name()(
-            port in any_nullable_port(),
-            dnsname in any::<String>(),
-        ) -> Relay {
-            Relay::SingleHostName(port, dnsname)
-        }
-    }
-
-    prop_compose! {
-        fn multi_host_name()(
-            dnsname in any::<String>(),
-        ) -> Relay {
-            Relay::MultiHostName(dnsname)
-        }
-    }
-
-    fn any_relay() -> BoxedStrategy<Relay> {
-        prop_oneof![single_host_addr(), single_host_name(), multi_host_name(),].boxed()
-    }
-
-    prop_compose! {
-        pub fn any_pool_params()(
-            id in any_pool_id(),
-            vrf in any::<[u8; 32]>(),
-            pledge in any::<u64>(),
-            cost in any::<u64>(),
-            margin in 0..100u64,
-            reward_account in any::<[u8; 28]>(),
-            owners in any::<Vec<[u8; 28]>>(),
-            relays in proptest::collection::vec(any_relay(), 0..10),
-        ) -> PoolParams {
-            PoolParams {
-                id,
-                vrf: Hash::new(vrf),
-                pledge,
-                cost,
-                margin: RationalNumber { numerator: margin, denominator: 100 },
-                reward_account: [&[0xF0], &reward_account[..]].concat().into(),
-                owners: owners.into_iter().map(|h| h.into()).collect::<Vec<Hash<28>>>().into(),
-                relays,
-                metadata: Nullable::Null,
-            }
-        }
-    }
+    use amaru_kernel::{prop_cbor_roundtrip, tests::any_pool_params};
+    use proptest::{collection, collection::vec, prelude::*};
 
     pub fn any_future_params(epoch: Epoch) -> impl Strategy<Value = (Option<PoolParams>, Epoch)> {
         prop_oneof![
